refactor: extract setup dependency installer
